app: factor out bpf-state.ts
So far everything except parser has been stuffed into app.ts
It's becoming big enough to be confusing.

Also a clean separation between the UI and it's input is necessary for
long-term maintainability. This is a step in that direction.
